+++ benchmarks/benchmark_test.go
@@ -1,0 +1,9 @@
+package benchmarks
+
+import "testing"
+
+func BenchmarkPlannerPlaceholder(b *testing.B) {
+	for b.Loop() {
+		_ = "pgsync"
+	}
+}